+++ packages/mirror-media-next/components/shared/article-list-item.js
@@ -1,7 +1,6 @@
 import styled from 'styled-components'
-import { transformTimeDataIntoSlashFormat } from '../utils'
 
-import CustomNextImage from './custom-next-image'
+import CustomNextImage from '../custom-next-image'
 
 const ItemWrapper = styled.a`
   display: block;
@@ -9,16 +8,15 @@ const ItemWrapper = styled.a`
   width: 100%;
   margin: 0 auto;
   font-size: 18px;
-  ${({ theme }) => theme.breakpoint.xl} {
-    border-radius: 6px;
-    overflow: hidden;
-  }
 `
 
 const ImageContainer = styled.div`
   position: relative;
   width: 100%;
   height: 214px;
+  ${({ theme }) => theme.breakpoint.xl} {
+    height: 147px;
+  }
 `
 
 const ItemSection = styled.div`
@@ -48,17 +46,17 @@ const ItemSection = styled.div`
 `
 
 const ItemDetail = styled.div`
-  padding: 20px 20px 36px 20px;
-  background-color: #eee;
+  margin: 20px 20px 36px 20px;
   ${({ theme }) => theme.breakpoint.md} {
-    padding: 20px;
+    margin: 20px;
   }
   ${({ theme }) => theme.breakpoint.xl} {
-    padding: 16px 24px 20px;
+    margin: 8px 8px 40px 8px;
   }
 `
 
 const ItemTitle = styled.div`
+  color: #054f77;
   line-height: 25px;
   display: -webkit-box;
   -webkit-line-clamp: 3;
@@ -69,9 +67,7 @@ const ItemTitle = styled.div`
     height: 75px;
   }
   ${({ theme }) => theme.breakpoint.xl} {
-    line-height: 30px;
-    font-size: 20px;
-    height: 90px;
+    font-size: 18px;
   }
 `
 
@@ -89,16 +85,9 @@ const ItemBrief = styled.div`
   }
 
   ${({ theme }) => theme.breakpoint.xl} {
-    margin-top: 8px;
-    -webkit-line-clamp: 3;
+    margin-top: 20px;
+    -webkit-line-clamp: 4;
   }
-`
-
-const ItemDate = styled.div`
-  margin-top: 12px;
-  color: #9b9b9b;
-  font-size: 14px;
-  line-height: 1.4;
 `
 
 /**
@@ -107,7 +96,7 @@ const ItemDate = styled.div`
  * @param {import('../type/category').CategorySection} [props.section]
  * @returns {React.ReactElement}
  */
-export default function PremiumArticleListItem({ item, section }) {
+export default function ArticleListItem({ item, section }) {
   const itemSection =
     section || item.sections.find((section) => section.slug !== 'member')
 
@@ -124,9 +113,6 @@ export default function PremiumArticleListItem({ item, section }) {
       <ItemDetail>
         <ItemTitle>{item.title}</ItemTitle>
         <ItemBrief>{item.brief?.blocks[0]?.text}</ItemBrief>
-        <ItemDate>
-          {transformTimeDataIntoSlashFormat(item.publishedDate)}
-        </ItemDate>
       </ItemDetail>
     </ItemWrapper>
   )
