+++ packages/mirror-media-next/pages/category/[slug].js
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import client from '../../apollo/apollo-client'
 import { fetchPosts } from '../../apollo/query/posts'
 import { fetchCategorySections } from '../../apollo/query/categroies'
-import CategoryArticles from '../../components/category-articles'
+import CategoryArticles from '../../components/category/category-articles'
 import { GCP_PROJECT_ID } from '../../config/index.mjs'
 
 const CategoryContainer = styled.main`
