+++ packages/mirror-media-next/pages/section/[slug].js
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import client from '../../apollo/apollo-client'
 import { fetchPosts } from '../../apollo/query/posts'
 import { fetchSection } from '../../apollo/query/sections'
-import SectionArticles from '../../components/section-articles'
+import SectionArticles from '../../components/shared/section-articles'
 import { GCP_PROJECT_ID } from '../../config/index.mjs'
 
 const SectionContainer = styled.main`
