+++ packages/mirror-media-next/pages/tag/[slug].js
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import client from '../../apollo/apollo-client'
 import { fetchPosts } from '../../apollo/query/posts'
 import { fetchTag } from '../../apollo/query/tags'
-import TagArticles from '../../components/tag-articles'
+import TagArticles from '../../components/tag/tag-articles'
 import { GCP_PROJECT_ID } from '../../config/index.mjs'
 
 const TagContainer = styled.main`
